moved gmsh2triangle to MeshTools; added docstring
